+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 
 * `LinearLayout` now implements `ViewGroup`
 * `LinearLayout::inner{_mut}`
+* `Orientation::compute_offset`
+
+## Changed
+
+* `Orientation::place` now returns the bounds of the placed View.
 
 0.3.2 (2023-08-23)
 ==================